勉強時間のデータ型をdoubleに変更

+++ src/database/migrations/2022_10_06_111531_create_content_records.php
@@ -16,7 +16,7 @@ class CreateContentRecords extends Migration
         Schema::create('content_records', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->date('date');
-            $table->integer('study_hour');
+            $table->double('study_hour');
             $table->integer('user_id');
             $table->integer('learning_content_id');
             $table->timestamps();
